more import fixes; moves bots
