+++ lib/elfpy/elfpy/tests/pricing_models/test_calc_in_given_out.py
@@ -6,15 +6,6 @@ import logging
 import unittest
 from typing import TYPE_CHECKING
 
-from calc_test_dataclasses import (
-    CalcInGivenOutFailureTestCase,
-    CalcInGivenOutSuccessByModelTestResult,
-    CalcInGivenOutSuccessTestCase,
-    CalcInGivenOutSuccessTestResult,
-)
-from fixedpointmath import FixedPoint
-from fixedpointmath import errors as fperrors
-
 import elfpy.time as time
 import elfpy.types as types
 import elfpy.utils.logs as log_utils
@@ -24,6 +15,14 @@ from elfpy.markets.hyperdrive import (
     YieldspacePricingModel,
     hyperdrive_actions,
 )
+from elfpy.tests.pricing_models.calc_test_dataclasses import (
+    CalcInGivenOutFailureTestCase,
+    CalcInGivenOutSuccessByModelTestResult,
+    CalcInGivenOutSuccessTestCase,
+    CalcInGivenOutSuccessTestResult,
+)
+from fixedpointmath import FixedPoint
+from fixedpointmath import errors as fperrors
 
 if TYPE_CHECKING:
     from elfpy.markets.base import BasePricingModel
