+++ src/eth_bots/core/get_agent_accounts.py
@@ -13,9 +13,9 @@ from numpy.random._generator import Generator as NumpyGenerator
 from web3 import Web3
 from web3.contract.contract import Contract
 
-from elfpy import eth
-from elfpy.eth.accounts import EthAgent
 from eth_bots.core import AgentConfig
+from src import eth
+from src.eth.accounts import EthAgent
 
 # pylint: disable=too-many-locals
