+++ src/eth_bots/core/trade_loop.py
@@ -9,8 +9,8 @@ from web3 import Web3
 from web3.contract.contract import Contract
 from web3.types import RPCEndpoint
 
-from elfpy.eth.accounts import EthAgent
 from eth_bots.core.execute_agent_trades import async_execute_agent_trades
+from src.eth.accounts import EthAgent
 
 
 def trade_if_new_block(
